fixed expression but with multiply vector by scalar

+++ python/maya_math_nodes/expression_info.py
@@ -55,7 +55,7 @@ FUNCTIONS = {
         'name': 'math_Multiply{0}',
         'attributes': ['input1', 'input2'],
         'types': ['double', 'double3', 'double4', 'doubleAngle', 'doubleLinear', 'int', 'matrix'],
-        'mixed_types': {'double': ['int'], 'double3': ['int', 'matrix'], 'doubleAngle': ['int'], 'doubleLinear': ['int']}
+        'mixed_types': {'double': ['int'], 'double3': ['double', 'matrix'], 'doubleAngle': ['int'], 'doubleLinear': ['int']}
     },
     '-': {
         'name': 'math_Subtract{0}',
